+++ api-server/src/main/java/com/kuddy/apiserver/spot/service/RecommendApiService.java
@@ -9,9 +9,11 @@ import com.kuddy.common.response.StatusResponse;
 import com.kuddy.common.spot.domain.Spot;
 import com.kuddy.common.spot.exception.NoSpotNearbyException;
 import com.kuddy.common.spot.exception.SpotRecommendApiException;
-import com.kuddy.common.spot.exception.TourApiExeption;
 import com.kuddy.common.spot.repository.SpotRepository;
+
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.ResponseEntity;
@@ -25,37 +27,39 @@ import java.util.List;
 @Service
 @Transactional
 @RequiredArgsConstructor
+@Slf4j
 public class RecommendApiService {
     private final RestTemplate restTemplate = new RestTemplate();
     private final SpotRepository spotRepository;
 
-    private static final String RECOMMEND_SERVER_URL = "http://localhost:8000/recommend/";
+    private static final String RECOMMEND_SERVER_URL = "https://api.kuddy.co.kr/recommend/";
 
-    public List<SpotResDto> getFiveSpotsByDistance(Member member, Long contentId, String mapX, String mapY) {
+    @Cacheable(value="spotRecommend",cacheManager="recommendationCacheManager")
+    public List<SpotResDto> getFiveSpotsByDistance(Member member, Long contentId) {
+        String url = RECOMMEND_SERVER_URL + contentId + "/" + member.getId();
+
+        List<Long> recommendedSpotIds = fetchRecommendedSpotIds(url);
+
+        if (recommendedSpotIds == null || recommendedSpotIds.isEmpty()) {
+            throw new NoSpotNearbyException();
+        }
+
+        List<Spot> spotList = spotRepository.findAllById(recommendedSpotIds);
+
+        return changeListToDto(spotList);
+    }
+
+    private List<Long> fetchRecommendedSpotIds(String url) {
         try {
-            // Recommend 서버 URL에서 장소 ID 가져오기
-            String url = RECOMMEND_SERVER_URL + contentId + "/" + member.getId() + "?x=" + mapX + "&y=" + mapY;
-
             ResponseEntity<List<Long>> responseEntity = restTemplate.exchange(
                     url,
                     HttpMethod.GET,
                     null,
                     new ParameterizedTypeReference<List<Long>>() {}
             );
-
-            List<Long> recommendedSpotIds = responseEntity.getBody();
-
-            if (recommendedSpotIds == null || recommendedSpotIds.isEmpty()) {
-                throw new NoSpotNearbyException();
-            }
-
-            // 가져온 장소 ID를 사용하여 상세 정보 조회
-            List<Spot> spotList = spotRepository.findAllById(recommendedSpotIds);
-
-            return changeListToDto(spotList);
-
+            return responseEntity.getBody();
         } catch (Exception e) {
-            e.printStackTrace();
+            log.error("Error fetching recommended spot IDs", e);
             throw new SpotRecommendApiException();
         }
     }
